% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{all_panel}
\alias{all_panel}
\title{Generate a Panel of Dot Plots for ALL Markers}
\usage{
all_panel(seu, color = "vmR_pred", cols)
}
\arguments{
\item{seu}{A Seurat object containing ADT data.}

\item{color}{A string representing the name of the feature to use for coloring the dots. Defaults to \code{"vmR_pred"}.}

\item{cols}{A vector of colors to use for the coloring the dots. Can be used for both continuous and categorical color scales.}
}
\value{
A patchwork object representing a grid of dot plots.
}
\description{
This function creates a panel of dot plots for Acute Myeloid Leukemia (ALL) marker expression, using ADT data from a Seurat object. Each plot compares pairs of AML-related surface markers, and the plots are arranged in a grid. The color of the points can be customized based on a specified feature.
}
\examples{
# Example usage
aml_panel(seu = my_seurat_object, color = "vmR_pred", cols = my_colors)

}
