% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\name{check_input}
\alias{check_input}
\title{Check Input Validity}
\usage{
check_input(input = NULL, name = NULL, valid = NULL)
}
\arguments{
\item{input}{The input object to validate.}

\item{name}{The name of the parameter being checked.}

\item{valid}{A vector of valid types/classes for the input.}
}
\description{
This is an internal helper function that checks the validity of input parameters.
}
\keyword{internal}
