% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_LSI}
\alias{cluster_LSI}
\title{Cluster LSI (Compatible with Monocle3 and Seurat)}
\usage{
cluster_LSI(
  object,
  k = 20,
  weight = FALSE,
  num_iter = 1,
  resolution = NULL,
  random_seed = 2020,
  verbose = TRUE,
  partition_qval = 0.05
)
}
\arguments{
\item{object}{The `cell_data_set` or Seurat object upon which to perform this operation.}

\item{k}{Integer number of nearest neighbors to use when creating the k nearest neighbor graph for Leiden clustering. Default is 20.}

\item{weight}{A logical argument to determine whether or not to use Jaccard coefficients for two nearest neighbors (based on the overlapping of their kNN) as the weight used for Louvain clustering. Default is FALSE.}

\item{num_iter}{Integer number of iterations used for Leiden clustering. Default is 1.}

\item{resolution}{Parameter that controls the resolution of clustering. If NULL (Default), the parameter is determined automatically.}

\item{random_seed}{The seed used by the random number generator. Default is 2020.}

\item{verbose}{A logical flag to determine whether or not to print the run details.}

\item{partition_qval}{Numeric, the q-value cutoff to determine when to partition. Default is 0.05.}
}
\description{
This function extracts clustering from the last iteration of LSI (see \code{iterative_LSI}) in a single-cell experiment. It uses Leiden clustering and computes partitions. Compatible with both Monocle3 `cell_data_set` and Seurat objects.
}
\references{
Granja, J. M.et al. (2019). Single-cell multiomic analysis identifies regulatory programs in mixed-phenotype acute leukemia. Nature Biotechnology, 37(12), 1458–1465.
}
\keyword{internal}
