% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract Data Using Specified Features (Compatible with Seurat Objects)}
\usage{
extract_data(
  query,
  subject,
  annotation = "row.names",
  verbose = TRUE,
  fill_output = TRUE,
  duplicate_hits = "max.disp",
  ignore_strand = TRUE
)
}
\arguments{
\item{query}{A vector of feature names (for annotation-based search) or a `GRanges` object (for range-based search).}

\item{subject}{A `SummarizedExperiment` or Seurat object from which to extract the data.}

\item{annotation}{Character, specifies the target feature column in `subject` to search for features. Default is "row.names".}

\item{verbose}{Logical, whether to print verbose output. Default is `TRUE`.}

\item{fill_output}{Logical, whether to return the `subject` object filled with zeros for features in `query` that are not found in `subject`. Default is `TRUE`.}

\item{duplicate_hits}{Character, specifies how to resolve multiple hits in range-based searches. Options are: "max.mean", "max.var", "max.disp", "min.mean", "min.var", "min.disp". Default is "max.disp".}

\item{ignore_strand}{Logical, whether to ignore strand information in range-based searches. Default is `TRUE`.}
}
\value{
A list containing:
- `mat`: A matrix of the extracted data.
- `overlap`: The ratio of features found in the `subject` compared to the `query`.
- `not_found`: Features not found in the `subject`.
}
\description{
This function extracts data from a `SummarizedExperiment`-like object or a Seurat object using features specified in the `query`. Features can be identified by either name (annotation-based) or genomic ranges (range-based).
}
\keyword{internal}
