% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{files3_prep}
\alias{files3_prep}
\title{Find Cell Ranger Data Files}
\usage{
files3_prep(folder)
}
\arguments{
\item{folder}{Path to the Cell Ranger output folder.}
}
\value{
The path to the folder containing the data files. If no files are found, returns \code{NULL}.
}
\description{
Identifies the correct path for Cell Ranger output files (matrix.mtx, barcodes.tsv, genes.tsv or features.tsv), possibly in compressed or uncompressed forms.
}
\examples{
\dontrun{
data_folder <- files3_prep("/path/to/cellranger/output")
}
}
