% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flscuts.R
\docType{package}
\name{flscuts-package}
\alias{flscuts}
\alias{flscuts-package}
\title{flscuts: Furlan Lab Shortcuts}
\description{
Not much more to say...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://furlan-lab.github.io/flscuts/}
}

}
\author{
\strong{Maintainer}: Scott Furlan \email{scottfurlan@gmail.com} (\href{https://orcid.org/0000-0003-3007-9840}{ORCID})

}
