% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GMT_to_list}
\alias{GMT_to_list}
\title{Convert a GMT File to a List of Gene Sets}
\usage{
GMT_to_list(gmt_file)
}
\arguments{
\item{gmt_file}{A character string specifying the path to the GMT file.}
}
\value{
A named list where each element is a vector of gene symbols for a gene set.
}
\description{
This function reads a Gene Matrix Transposed (GMT) file and converts it into a list where each element corresponds to a gene set. The names of the list elements are the gene set names, and the contents are vectors of genes belonging to each set.
}
\details{
The GMT file format is commonly used for storing gene sets, such as those provided by MSigDB. Each line in a GMT file corresponds to a gene set and is formatted as follows:
\enumerate{
\item Gene set name
\item Description (can be empty)
\item Gene symbols (tab-separated)
}
This function reads the GMT file line by line, splits each line by tabs, and extracts the gene set name and the list of genes.
}
\examples{
\dontrun{
# Convert a GMT file to a list of gene sets
gene_sets <- GMT_to_list("path/to/your/file.gmt")
}
}
\keyword{internal}
