% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_geneset}
\alias{plot_geneset}
\title{Plot Geneset Scores on Cell Embeddings}
\usage{
plot_geneset(
  cds,
  marker_set,
  name,
  fData_col = "gene_short_name",
  method = c("totals", "corrected"),
  reduction_method = "UMAP",
  cell_size = 0.5
)
}
\arguments{
\item{cds}{A \code{cell_data_set} object from the \code{monocle3} package.}

\item{marker_set}{A character vector of gene identifiers corresponding to the genes in the geneset.}

\item{name}{A character string representing the name of the geneset (used as a label).}

\item{fData_col}{A character string specifying the feature data column in \code{rowData(cds)} that contains the gene identifiers in \code{marker_set}. Default is \code{"gene_short_name"}.}

\item{method}{A character string specifying the method to calculate geneset scores. Either \code{"totals"} or \code{"corrected"}. Default is \code{"totals"}.}

\item{reduction_method}{A character string specifying the dimensionality reduction method to use for plotting. Default is \code{"UMAP"}.}

\item{cell_size}{Numeric value specifying the size of the points in the plot. Default is \code{0.5}.}
}
\value{
A ggplot object representing the cells colored by geneset scores.
}
\description{
Plots geneset scores for each single cell on a dimensionality reduction embedding (e.g., UMAP or t-SNE). The geneset scores are calculated per cell based on a set of genes.

When using method \code{"totals"}, the sum of the size-factor corrected, log-normalized gene expression for the given set of genes is computed. When using method \code{"corrected"}, the geneset scores are corrected using 100 random genes with similar expression levels.
}
\examples{
# Assuming `cds` is a cell_data_set object, and `genes` is a vector of gene identifiers
# plot_geneset(cds, marker_set = genes, name = "MyGeneset")

}
\references{
Puram, S. V. et al. (2017). Single-Cell Transcriptomic Analysis of Primary and Metastatic Tumor Ecosystems in Head and Neck Cancer. \emph{Cell}, 171(7), 1611–1624.e24.
}
