% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\name{project_data2}
\alias{project_data2}
\title{Project data from one object into the UMAP embedding a single cell object}
\usage{
project_data2(
  projector = NULL,
  projectee = NULL,
  ncells_coembedding = 5000,
  scale = F,
  reduced_dim = "LSI",
  embedding = "UMAP",
  make_pseudo_single_cells = FALSE,
  features = c("annotation-based", "range-based"),
  n = 250,
  verbose = TRUE,
  threads = 6,
  seed = 2020,
  force = F
)
}
\arguments{
\item{projector}{cell_data_set object with a reduced dimension matrix (currently LSI supported) as specified in
reduced_dim argument and a model used to create a low dimensional embedding}

\item{projectee}{a SummarizedExperiment type object (cell_data_set currently supported) to be projected using
the models contained in the projector}

\item{ncells_coembedding}{number of cells in the projector to use in in the co-embedding with simulated single cells; default is 5000, will automatically
default to the total number of cells in the projector if less than this value.}

\item{reduced_dim}{A string specifying the reducedDim (currently LSI and PCA supported).}

\item{embedding}{A string specifying embedding type (currently UMAP supported).}

\item{make_pseudo_single_cells}{whether to make pseudo-single cells from the data in the projectee (set this to true for bulk data)}

\item{n}{An integer specifying the number of subsampled "pseudo single cells" per bulk sample.  Note this is only relevant if
make_pseudo_single_cells is TRUE}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can be set to FALSE for a cleaner output.}

\item{threads}{The number of threads used for parallel execution}
}
\description{
Adapted from: ArchR: An integrative and scalable software package for single-cell chromatin accessibility analysis
Jeffrey M. Granja, M. Ryan Corces, Sarah E. Pierce, S. Tansu Bagdatli, Hani Choudhry, Howard Y. Chang, William J. Greenleaf
doi: https://doi.org/10.1101/2020.04.28.066498
}
