% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\name{safe_subset}
\alias{safe_subset}
\title{Subset Matrix Safely}
\usage{
safe_subset(mat = NULL, subsetRows = NULL, subsetCols = NULL)
}
\arguments{
\item{mat}{A matrix to subset.}

\item{subsetRows}{Character vector specifying row names to subset.}

\item{subsetCols}{Character vector specifying column names to subset.}
}
\value{
A subsetted matrix, potentially filled with zeros.
}
\description{
Safely subsets a matrix, filling in missing rows or columns with zeros if necessary.
}
\keyword{internal}
