% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seurat_to_monocle3}
\alias{seurat_to_monocle3}
\title{Convert Seurat Object to Monocle3 Cell Data Set}
\usage{
seurat_to_monocle3(
  seurat_obj,
  assay_name = "RNA",
  seurat_reduction = "umap",
  monocle_reduction = "UMAP"
)
}
\arguments{
\item{seurat_obj}{A Seurat object to convert.}

\item{assay_name}{A character string specifying the assay to use from the Seurat object. Default is \code{"RNA"}.}

\item{seurat_reduction}{The name of the reduction in Seurat to transfer (e.g., \code{"umap"}). Use \code{NULL} to skip transferring reductions. Default is \code{"umap"}.}

\item{monocle_reduction}{The name to assign to the reduction in the Monocle3 object (e.g., \code{"UMAP"}). Use \code{NULL} to skip transferring reductions. Default is \code{"UMAP"}.}
}
\value{
A Monocle3 \code{cell_data_set} object.
}
\description{
Converts a Seurat object into a Monocle3 \code{cell_data_set} object. Optionally, dimensionality reduction embeddings can be transferred.
}
\details{
This function converts a Seurat object into a Monocle3 \code{cell_data_set} object by extracting the counts data, cell metadata, and gene metadata. It can also transfer dimensionality reduction embeddings from the Seurat object to the Monocle3 object if specified.
}
\examples{
\dontrun{
# Assuming 'seurat_obj' is a Seurat object
cds <- seurat_to_monocle3(
  seurat_obj,
  assay_name = "RNA",
  seurat_reduction = "umap",
  monocle_reduction = "UMAP"
)
}
}
