% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{tf_idf_transform}
\alias{tf_idf_transform}
\title{Performs TF-IDF transformation on a cell_data_set}
\usage{
tf_idf_transform(input, method = 1, verbose = TRUE)
}
\arguments{
\item{cds_list}{Input cell_data_set object or sparse matrix.}
}
\description{
Just like it sounds.
}
\keyword{internal}
