% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{sparseRowVariances}
\alias{sparseRowVariances}
\title{Sparse Row Variances}
\usage{
sparseRowVariances(m)
}
\arguments{
\item{m}{A sparse matrix of class \code{dgCMatrix}.}
}
\value{
A numeric vector of row variances.
}
\description{
Computes the variances of rows in a sparse matrix efficiently.
}
