% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{tf_idf_transform}
\alias{tf_idf_transform}
\title{TF-IDF Transformation}
\usage{
tf_idf_transform(mat)
}
\arguments{
\item{mat}{A sparse matrix.}
}
\value{
A sparse matrix after TF-IDF transformation.
}
\description{
Performs TF-IDF transformation on a matrix.
}
